Timeline: Add functions to find "best" and parent index in a model
This is useful if you have a timestamp and are looking for a range that
is "close" to that timestamp. It's hard to determine a range that
actually covers the timestamp, but the "best" index gives a pretty good
starting point for a subsequent linear search.

In order to determine if there cannot be any earlier range that touches
a given timestamp we need to check the parent range. Add a function that
retrieves it.

Change-Id: I2a3cad006ff8449620c56899ab8166d1b62d6747
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>