ProjectExplorer: Further orthogonalize Target/Tools
This introduces a 'TargetRunner' base that it meant as a base
class for classes like the current AndroidRunner, IosRunner,
WinRTRunnerHelper, similar to the existing 'ToolRunner' class
taking care of the tool bits.

This makes a RunControl only responsible for state management
(start/stop control), application output and general user
visible display tasks, delegation most of the real work either
to the the target or the tool.

The usage pattern for creating a run control is now

RunControl *RunControlFactory::create(...)
{
    if (barToolOnFooTargetIsSupported()) {
        auto runControl = new RunControl(runConfig, runMode);
        (void) new FooTargetRunner(runControl);
        (void) new BarToolRunner(runControl);
        return runControl;
    }
    ...
}

The pattern can possibly be simplified later, for now the goal
is to make it uniformly used.

As first use, "split" SimpleRunControl into a SimpleToolRunner
(which happens to be all its previous non-(Base)RunControl bits)
and the run control bits (its RunControl base).

This will make the SimpleRun*Control* class as such vanish again,
keep it for now to give downstream uses (Qnx/Boot2Qt) some time
window to adapt.

Change-Id: Ifafed4d0577466186280ac5014f4ad9eec78c56c
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>