QmlDesigner: Handle missing type information
We always had the problem that the design mode might become
active while type information is still being loaded.
This causes an error message and the user had to trigger
a reload of the QML document.
Due to changes in the project management Qt Creator
starts reading the type information, once the first QML file
is opened.
This triggers the error far more frequently.

This patch fixes the issue by triggering a reaload after 1s,
if type information has not been read completely.
This fixes the issue.

Change-Id: I23ad011294044690fd582e808227c3ccde8bda54
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>