Qmake: Fix workdirectory for qmake projects
The rootProjectNode was not set when the desktopqmakerunconfiguration tried
to use it to figure out the default work directory. Move this code over to
using rootProFile() instead of rootProjectNode(), which is the right thing
to do: rootProFile() is the tree of objects used to parse the qmake project
now.

Task-number: QTCREATORBUG-17728
Change-Id: I76167ba2bbd395d58af7d3bfd50f33c4e37338db
Reviewed-by: hjk <hjk@qt.io>