Utils: Keep correct palette for dark themers after suspend
Sleep, hibernation and user sign-in can cause a ThemeChange event. That
event resets the application palette, discarding a palette that may have
been previously set.

Dark themes in Qt Creator want to set and keep their custom application
palettes. So, this change sets the custom application palette on each
ThemeChange event that is send to the main window.

Task-number: QTCREATORBUG-14929
Change-Id: Ic9fb0111cfa0e8171b819d687f280c3db6cc8f2c
Reviewed-by: David Schulz <david.schulz@qt.io>