QmlDesigner: Crash fix
This is a nasty one. The text item that
triggers the crash is part of a delegate and therefore,
does not have an item in the scene.
The rewriter is now setting the id again, when reading the file after
the Rectangle is removed. we get a crash.
Since delegates are currently not shown in the navigator, this crash
could not happen before we introduced the integrated text editor.


Task-number: QTCREATORBUG-17970
Change-Id: Ica16dec5bb2f852324de179c4a2b4d43dcf27aa9
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>