Qmake: Update QmakeRunConfiguration in constructor
Update QmakeRunConfiguration's aspects in the constructor. This is necessary
since this RC is only created after the project was parsed for the first time,
so the logic tied to the pro-file parsing updates did not work the first time
round.

Also move the logic from the UI into the RC itself.

Task-number: QTCREATORBUG-18059
Change-Id: Ifa6eeade549ef33aae5c825238fc8a1d76604195
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>