Highlighter: Fix freezing highlighter for complex context switches
Exclamation marks in context attributes are used to split the attribute
into a number of orders and an identifier. This identifier is the
context that have to be set after executing the orders. Ignoring this
context identifier could lead to an endless loop if used inside a
default context with a Rule that has set the lookAhead attribute.

Task-number: QTCREATORBUG-14611
Change-Id: I5992fa47ed2e353cbf7882bc772fdbee8f7f41f1
Reviewed-by: Christian Stenger <christian.stenger@qt.io>