QmlDesigner: Reimplement navigator model
This patch gets rid of of QStandardItemModel.
Instead we implement a simple tree model without
any explicit structure. The model simply maps
the generic Model to a tree.

The advantage of the new model is that is does not use
a shadow data structure (QStandardItemModel) anymore.
The original model is always directly mapped to the tree model.
This makes initialization a lot faster and and there cannot
be any synchronization issues anymore.

Change-Id: I0e71ff2d4067f29a4c25c78ad3d626a5daf206a4
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>