RemoteLinux: Don't force value for DISPLAY in base environment
When set on nVidia Jetson TX-1, it leads to applications failing.

To allow previously created projects to work, add DISPLAY=:0.0 to user
changes to the base environment the first time the project is loaded. On
new projects for devices using X11 the user now has to manually set the
DISPLAY variable.

Task-number: QTBUG-60665
Change-Id: Iecc192fbad81ad5cbbbcabce6aeb28c3f501d022
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>