QmlProfiler: Fix flame graph context menu test
The context menu event can be generated multiple times within one loop
of QTRY_VERIFY. The result is that multiple showFullRange() signals can
be generated before we check again for spy.count() == 1. Thus, the check
never succeeds and the number of signals keeps growing.

We connect the showFullRange() signal to the model manager in order to
get a more realistic test setup. This way the action that generates the
showFullRange() signal is disabled in any further context menus, just as
it is supposed to be. In addition we can now check for the manager to
actually show the full range.

Change-Id: I5e13c2666ce1a15c7a5fad57affd4274d9656656
Reviewed-by: Christian Stenger <christian.stenger@qt.io>