Squish: Fix switching to build or run settings
If too many kits are present and/or the screen resolution is
too low to have QC having the correct minimum windows size
Squish fails to click on the items as it might be necessary
to scroll to them to have them really (user-) visible.
There is an obvious difference between 'visible' for Squish
and 'visible' for the interacting user.

Change-Id: Id658a323fcf2ae3882b225c6ce9383ed2c876e81
Reviewed-by: Robert Loehning <robert.loehning@qt.io>