Resources: Read qrc file contents from file if that is readable
Retrieve the qrc file contents from the file if that is readable. This
is now necessary since the ResourceTopLevelNode now contains the QRC
file contents itself so the QML code model has access to it, even if
a qrc file is actually used.

This fixes one part of QTCREATORBUG-17930

Task-number: QTCREATORBUG-17930
Change-Id: I0e8cc8828062f079d8634882a768ca60331a0e16
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>