QmlDesigner: Always add the iso icons to the qrc mapping
Since the qrc scanning is now asynchronous the qrc mapping
is not always reliable. Especially the first time when a new
.qrc file is added.
To keep at least the iso icon tool working reliable we always
assume an iso-icons directory.
Since the mapping qrc to file system is one to many anyway, this is
not breaking anything.

Task-number: QTCREATORBUG-18225
Change-Id: Ibb6d77960bd643f8c1f37bebc0c0e400588d0429
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>