CodeAssist: Fix leaking proposal model
The model is not owned by the proposal.

For corner cases, we leaked here already for ages. Starting with

  commit fab4dd068ed51f612e7beaf0985bcbe4411d2f01
  CodeAssist: Fix auto completion if function signature is shown

leaks happened also for a hot path: code completion results that does
not match any so far typed prefix. Due to auto completion turned on,
this triggers frequently when typing text.

Task-number: QTCREATORBUG-17752
Task-number: QTCREATORBUG-18326
Change-Id: I9df177723b31cf3df36e73ef1e7920d9aa5e16e5
Reviewed-by: Marco Bubke <marco.bubke@qt.io>