QmlDesigner: Fix combobox for font family
The editable combobox had some serious issues.

I changed the signals we react to and the properties
we use.

We react on activated to change to fonts in the model. For some
reason using currentText does not work reliable.
We have to retrieve the current etxt manually using
textAt().

We also react to accepted to changes in the text edit.
Here we simply use editText.

We also react to selection changes and change
the font.

Change-Id: I99abd7609f8b64ef446ce154aed0c2a61dfa00f9
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>