AutoTest: Improve finding executables for cmake based projects
Depending on the structure of a project we might end up inside a
library and will not find the correct executable for running a
test. This patch improves finding the correct executable, but
depending on the complexity of the project it will not be able
to fix all circumstances.

Task-number: QTCREATORBUG-17882
Change-Id: I847bb40426bc1b874c2dc6a2a6880b11fe048bda
Reviewed-by: David Schulz <david.schulz@qt.io>