QmlDesigner: Refine selection
This is a major overhaul of the slection logic.
We do not change to the move tool automatically once
the curser enters the selected item. This allows selection
of items that are at the same position as the current selected item.

The selected item can still be moved if there is no pissible candidate for
selection. Also the border and gizmo always allows an item to be moved, if
the item is fully covered by another item.

Change-Id: Ic97a2bf23f33fcc5e209248aeb2f97df67bd23e1
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>