Fix the deployment target problem once and for all
Instead of hardcoding the value of LSMinimumSystemVersion in Info.plist
where we will always forget to update it (and since Qt Creator can be
built with multiple versions of Qt, it will almost certainly be wrong
for one of those versions), automatically determine the value from
QMAKE_MACOSX_DEPLOYMENT_TARGET, which is authoritative.

This should prevent mishaps where users try to run Qt Creator on older
OS versions and receive a crash dialog instead of the friendly "This
app requires a newer version of macOS" message that a properly set
LSMinimumSystemVersion value will induce.

The Qbs build is not affected by this problem, as Qbs automatically
handles the minimum deployment target even when a custom Info.plist is
specified.

The solution chosen in this patch was used instead of QMAKE_SUBSTITUTES
because the Qbs build also uses the Info.plist input file and it would
break the Qbs build if the quotes were to be escaped, since it would
become invalid XML.

Change-Id: I20625a2fae546c6597a567f28864b12917e8ac39
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>