Kit: Improve toolchain setup on auto-detected kit setup
The auto-detected Kit used by default in creator used to set up a
set of toolchains, that usually match up with each other (this relies
on the auto-detecion of toolchains to detect groups of compilers
in one go).

Then the Qt version selected (or better: The mkspec) will change the
C++ compiler if necessary.

Change this to actually pick a good C++ compiler and also update all
other toolchains to match the C++ compiler.

Task-number: QTCREATORBUG-18571
Change-Id: I95ddd9b83cf1642fba648919b76d3e3b7aa24c43
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>