QmlDesigner: Adjust margings for "move zone"
I introduced a zone around the boarder of each item where
it can always be dragged,
instead of selection another overlapping item.
The area is now a lot smaller, but the 4px seem enough for me
when testing it.
This zone is slightly to big. I keep the size of 20px at the top
so the label/gizmo is always dragable.

Change-Id: Ia1b8afb822b506f5322b952d5537eafc73b52ea2
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>