SCXML editor: allow eventless transitions
According to SCXML specification (https://www.w3.org/TR/scxml/), the
'event' attribute is not mandatory in 'transition' tags.
Also, as specified in section 3.12.1, eventless transitions are handled
differently, and will be taken whenever the 'cond' attribute evaluates
to 'true'.

This patch removes the automatic assignment of 'event' property in the
tag constructors, but restore it when a new transition is done in the
editor.  In this way, the user can delete the unneeded events without
them being restored when reloading the document.

Task-number: QTCREATORBUG-18345
Change-Id: I96d6860ad3ed58ab49e83d3328e88ef984f40fa4
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>