Support debug requests from Task Manager
If debugging was started from Task Manager Windows does
not follow the format given in the registry but always uses
"<debugger> -p <pid>".

By ignoring the -p parameter and using "debug" instead of
"wincrashevent" we can also support this use case from QtC's
debugger.

The message which is shown in this case will be adapted as
soon as this change hits master but with the current approach
the situation is greatly improved for 4.4.

Task-number: QTCREATORBUG-18194
Change-Id: I871a19f0cd68f61337d1e6c224ecc3c22a02c989
Reviewed-by: Robert Loehning <robert.loehning@qt.io>
Reviewed-by: hjk <hjk@qt.io>