QmlDesigner: Change selection in the release event
We already change the selection currently in the press event,
which leads wo weird behaivour if e.g. shift is pressed.
In this case the item is immediately unselected which is unexpected
and cannot be moved anymore.
Generally reacting to the event in the release event is the correct way
to do it.

Change-Id: Ib765f97f85e5dbeb6148d48f8640f1798a057b48
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>