ProjectExplorer: Workaround for displaying startup failures
Failures during RunControl construction currently get not to
the user as the Application output pane is not yet available.

In theory it should not happen, rather the condition should
be checked to disable running entirely, but in practice it
can happen e.g. in the BareMetal setup right now.

To avoid pre-release code restructuring, force a message box
in such cases.

Change-Id: I703c47b270d153afff34e08dad7833540a5983a2
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>