Reduce dependencies for sdktool to QtCore
Explicitly pull in the sources that are needed from utils, instead of
linking against the library.
Also generate another copy of app_version.h for sdktool, so
it is not necessary to run qmake on src/app/ before building sdktool.
For Qbs, it already depends on the separate app_version_header product.

Change-Id: I102c3a928f88a6035640166035c1bca29b126068
Reviewed-by: Jake Petroules <jake.petroules@qt.io>
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>