Help: Make restoring override cursor more robust
Some help backends are not too consistent in how they send started and
finished signals (QtWebEngine I'm looking at you), so make restoring the
override cursor more robust in the base. As long as the backend reports
that it is finished, guarantee that all previous override cursors are
restored.
Also ensure override cursors are restored in case the viewer is
destroyed.

Task-number: QTCREATORBUG-17758
Change-Id: Icb1cd1eabfb232135a391a550d8e6719a74189eb
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>